requirements.txt
google-generativeai>=0.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
pytest>=7.4.3